requirements.txt
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
transformers==4.56.2
sentence-transformers==2.2.2
fastapi
uvicorn
python-multipart
pillow
PyMuPDF
python-pptx
camelot-py[cv]
chromadb
langchain
openai
python-dotenv
